+++ R/create.R
@@ -15,26 +15,30 @@
 #' @return The path to the created object (invisibly).
 #' @name create
 #' @examples
-#' \dontshow{fs:::pkgdown_tmp(c("/tmp/filedd461e481b37", "/tmp/filedd46ff2c769"))}
-#' x <- file_create(file_temp())
-#' is_file(x)
+#' \dontshow{.old_wd <- setwd(tempdir())}
+#' file_create("foo")
+#' is_file("foo")
 #' # dir_create applied to the same path will fail
-#' try(dir_create(x))
+#' try(dir_create("foo"))
 #'
-#' x <- dir_create(file_temp())
-#' is_dir(x)
+#' dir_create("bar")
+#' is_dir("bar")
 #' # file_create applied to the same path will fail
-#' try(file_create(x))
+#' try(file_create("bar"))
+#'
+#' # Cleanup
+#' file_delete("foo")
+#' dir_delete("bar")
+#' \dontshow{setwd(.old_wd)}
 #' @export
 file_create <- function(path, mode = "u=rw,go=r") {
   assert_no_missing(path)
-
-  path <- path_expand(path)
-
   stopifnot(length(mode) == 1)
-  mode <- as_fs_perms(mode)
 
-  create_(path, mode)
+  mode <- as_fs_perms(mode)
+  new <- path_expand(path)
+
+  create_(new, mode)
   invisible(path_tidy(path))
 }
 
@@ -42,13 +46,12 @@ file_create <- function(path, mode = "u=rw,go=r") {
 #' @rdname create
 dir_create <- function(path, mode = "u=rwx,go=rx", recursive = TRUE) {
   assert_no_missing(path)
-
-  path <- path_expand(path)
-
   stopifnot(length(mode) == 1)
-  mode <- as_fs_perms(mode)
 
-  paths <- path_split(path)
+  mode <- as_fs_perms(mode)
+  new <- path_expand(path)
+
+  paths <- path_split(new)
   for (p in paths) {
     if (length(p) == 1 || !isTRUE(recursive)) {
       mkdir_(p, mode)
@@ -72,16 +75,15 @@ dir_create <- function(path, mode = "u=rwx,go=rx", recursive = TRUE) {
 link_create <- function(path, new_path, symbolic = TRUE) {
   assert_no_missing(path)
   assert_no_missing(new_path)
-
-  path <- path_expand(path)
-  new_path <- path_expand(new_path)
-
   stopifnot(length(path) == length(new_path))
 
+  old <- path_expand(path)
+  new <- path_expand(new_path)
+
   if (isTRUE(symbolic)) {
-    link_create_symbolic_(path, new_path)
+    link_create_symbolic_(old, new)
   } else {
-    link_create_hard_(path, new_path)
+    link_create_hard_(old, new)
   }
 
   invisible(path_tidy(new_path))