+++ R/is.R
@@ -8,40 +8,38 @@
 #' @importFrom stats setNames
 #' @export
 #' @examples
-#' \dontshow{fs:::pkgdown_tmp("/tmp/filedd46c01c6ad")}
-#' tmp <- dir_create(file_temp())
+#' \dontshow{.old_wd <- setwd(tempdir())}
+#' dir_create("d")
 #'
-#' file_create(path(tmp, "file.txt"))
-#' dir_create(path(tmp, "dir"))
-#' link_create(path(tmp, "file.txt"), path(tmp, "link"))
+#' file_create("d/file.txt")
+#' dir_create("d/dir")
+#' link_create(path(path_abs("d"), "file.txt"), "d/link")
 #'
-#' paths <- dir_ls(tmp)
+#' paths <- dir_ls("d")
 #' is_file(paths)
 #' is_dir(paths)
 #' is_link(paths)
+#'
+#' # Cleanup
+#' dir_delete("d")
+#' \dontshow{setwd(.old_wd)}
 is_file <- function(path) {
-  path <- path_expand(path)
-
   res <- file_info(path)
-  setNames(res$type == "file", res$path)
+  setNames(!is.na(res$type) & res$type == "file", res$path)
 }
 
 #' @rdname is_file
 #' @export
 is_dir <- function(path) {
-  path <- path_expand(path)
-
   res <- file_info(path)
-  setNames(res$type == "directory", res$path)
+  setNames(!is.na(res$type) & res$type == "directory", res$path)
 }
 
 #' @rdname is_file
 #' @export
 is_link <- function(path) {
-  path <- path_expand(path)
-
   res <- file_info(path)
-  setNames(res$type == "symlink", res$path)
+  setNames(!is.na(res$type) & res$type == "symlink", res$path)
 }
 
 #' Test if a path is an absolute path