+++ R/list.R
@@ -21,18 +21,30 @@
 #' @template fs
 #' @export
 #' @examples
+#' \dontshow{.old_wd <- setwd(tempdir())}
 #' dir_ls(R.home("share"), type = "directory")
 #'
+#' # Create a shorter link
 #' link_create(system.file(package = "base"), "base")
 #'
 #' dir_ls("base", recursive = TRUE, glob = "*.R")
+#'
+#' dir_map("base", identity)
+#'
+#' dir_walk("base", str)
+#'
+#' dir_info("base")
+#'
+#' # Cleanup
+#' link_delete("base")
+#' \dontshow{setwd(.old_wd)}
 dir_ls <- function(path = ".", all = FALSE, recursive = FALSE, type = "any",
                    glob = NULL, regexp = NULL, invert = FALSE, ...) {
   assert_no_missing(path)
 
-  path <- path_expand(path)
+  old <- path_expand(path)
 
-  files <- as.character(dir_map(path, identity, all, recursive, type))
+  files <- as.character(dir_map(old, identity, all, recursive, type))
 
   path_filter(files, glob, regexp, invert = invert,...)
 }
@@ -50,44 +62,33 @@ directory_entry_types <- c(
 
 #' @rdname dir_ls
 #' @param fun A function, taking one parameter, the current path entry.
-#' @examples
-#' dir_map("base", identity)
 #' @export
 dir_map <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any") {
   assert_no_missing(path)
 
-  path <- path_expand(path)
-
   type <- match.arg(type, names(directory_entry_types), several.ok = TRUE)
 
-  dir_map_(path, fun, all, sum(directory_entry_types[type]), recursive)
+  old <- path_expand(path)
+
+  dir_map_(old, fun, all, sum(directory_entry_types[type]), recursive)
 }
 
 #' @rdname dir_ls
-#' @examples
-#' dir_walk("base", str)
 #' @export
 dir_walk <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any") {
   assert_no_missing(path)
 
-  path <- path_expand(path)
+  old <- path_expand(path)
 
-  dir_map(path, fun, all, recursive, type)
+  dir_map(old, fun, all, recursive, type)
   invisible(path_tidy(path))
 }
 
 #' @rdname dir_ls
 #' @export
-#' @examples
-#' dir_info("base")
-#'
-#' # Cleanup
-#' link_delete("base")
 dir_info <- function(path = ".", all = FALSE, recursive = FALSE,
                      type = "any", regexp = NULL, glob = NULL, ...) {
   assert_no_missing(path)
-
-  path <- path_expand(path)
 
   file_info(dir_ls(path = path, all = all, recursive = recursive, type = type,
     regexp = regexp, glob = glob, ...))