+++ R/temp.R
@@ -12,13 +12,20 @@ env$temp_names <- character()
 #' `file_temp_pop()` can be used to explicitly remove an entry from the
 #' internal stack, however generally this is done instead by calling
 #' `file_temp()`.
+#'
+#' `path_temp()` constructs a path within the session temporary directory.
 #' @param pattern A character vector with the non-random portion of the name.
 #' @param tmp_dir The directory the file will be created in.
 #' @param ext The file extension of the temporary file.
+#' @param ... Additional paths appended to the temporary directory by `path()`.
 #' @template fs
 #' @export
 #' @examples
 #' \dontshow{file_temp_push("/tmp/filedd461c46df20")}
+#'
+#' path_temp()
+#' path_temp("does-not-exist")
+#'
 #' # default just passes the arguments to `tempfile()`
 #' file_temp()
 #'
@@ -33,8 +40,6 @@ env$temp_names <- character()
 file_temp <- function(pattern = "file", tmp_dir = tempdir(), ext = "") {
   assert_no_missing(tmp_dir)
 
-  tmp_dir <- path_expand(tmp_dir)
-
   path_tidy(file_temp_pop() %||% tempfile(pattern, tmp_dir, ext))
 }
 
@@ -43,9 +48,9 @@ file_temp <- function(pattern = "file", tmp_dir = tempdir(), ext = "") {
 file_temp_push <- function(path) {
   assert_no_missing(path)
 
-  path <- path_expand(path)
+  old <- path_expand(path)
 
-  env$temp_names <- c(env$temp_names, path)
+  env$temp_names <- c(env$temp_names, old)
 
   invisible(path_tidy(path))
 }
@@ -63,3 +68,10 @@ file_temp_pop <- function() {
   env$temp_names <- env$temp_names[-1]
   path_tidy(out)
 }
+
+#' @export
+#' @rdname file_temp
+path_temp <- function(...) {
+  path(tempdir(), ...)
+}
+