+++ src/RcppExports.cpp
@@ -87,12 +87,12 @@ BEGIN_RCPP
 END_RCPP
 }
 // chmod_
-void chmod_(CharacterVector path, mode_t mode);
+void chmod_(CharacterVector path, IntegerVector mode);
 RcppExport SEXP _fs_chmod_(SEXP pathSEXP, SEXP modeSEXP) {
 BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< CharacterVector >::type path(pathSEXP);
-    Rcpp::traits::input_parameter< mode_t >::type mode(modeSEXP);
+    Rcpp::traits::input_parameter< IntegerVector >::type mode(modeSEXP);
     chmod_(path, mode);
     return R_NilValue;
 END_RCPP
@@ -274,13 +274,25 @@ BEGIN_RCPP
 END_RCPP
 }
 // expand_
-CharacterVector expand_(CharacterVector path);
-RcppExport SEXP _fs_expand_(SEXP pathSEXP) {
+CharacterVector expand_(CharacterVector path, bool windows);
+RcppExport SEXP _fs_expand_(SEXP pathSEXP, SEXP windowsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< CharacterVector >::type path(pathSEXP);
-    rcpp_result_gen = Rcpp::wrap(expand_(path));
+    Rcpp::traits::input_parameter< bool >::type windows(windowsSEXP);
+    rcpp_result_gen = Rcpp::wrap(expand_(path, windows));
+    return rcpp_result_gen;
+END_RCPP
+}
+// tidy_
+CharacterVector tidy_(CharacterVector path);
+RcppExport SEXP _fs_tidy_(SEXP pathSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< CharacterVector >::type path(pathSEXP);
+    rcpp_result_gen = Rcpp::wrap(tidy_(path));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -310,7 +322,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_fs_readlink_", (DL_FUNC) &_fs_readlink_, 1},
     {"_fs_realize_", (DL_FUNC) &_fs_realize_, 1},
     {"_fs_path_", (DL_FUNC) &_fs_path_, 2},
-    {"_fs_expand_", (DL_FUNC) &_fs_expand_, 1},
+    {"_fs_expand_", (DL_FUNC) &_fs_expand_, 2},
+    {"_fs_tidy_", (DL_FUNC) &_fs_tidy_, 1},
     {NULL, NULL, 0}
 };